requirements.txt
google-generativeai>=0.3.0
Pillow>=9.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
tqdm>=4.65.0